Fix issue which prevents config file edits from applying
If a single Docker volume is shared between containers instantiated from
images built with the Dockerfile in this repo, one container's copy of
the config directory at /etc/keylime overrides the other containers'
copies. This is a problem when each Dockerfile for the respective KL
components makes its own changes to this directory. As such, this commit
moves the instructions which edit the config files to the base
Dockerfile which the verifier, registrar and tenant Dockerfiles
"inherit" from.