Remove delay before creating agent container
Commit 1ca3248 introduced logic to wait for the keylime_verifier and
keylime_registrar containers to be available before creating the
keylime_agent container. This was in an attempt to stop tcp connect
errors from being printed to the agent's logs before the registrar is
reachable. However, upon further investigation, these errors do not
appear when the containers are first created but rather when the Docker
daemon is restarted (e.g., after a system reboot or when a new GCE VM
is instantiated from a GCE image). As such, this commit reverts this
change.